Fit and simulate HMMs with custom decision models
This update has a lot because I forgot to commit earlier. Ability to
generate data from a given HMM, as well as fit it using Baum-Welch. The
DecisionModel class allows for user to specify flexible ways of deciding
which action to take, given the stimulus. Confirmed to work with basic
categorical emission model.